Implement github.com/hslatman/ipstore as storage layer
This commit adds the new github.com/hslatman/ipstore package as
a dependency. The library can be used to store `interface{}` types
in an in-memory store using a net.IP or net.IPNet as its key.

The previous implementation used github.com/hashicorp/go-immutable-radix,
which is a great generic library for (immutable) radix trees, but wasn't
a great fit for our current use case. We used a byte per bit to encode
the binary tree, which took a bit more processing and more key storage
space.