Add option to fail hard (or not) on API connection errors
When trying to connect to the CrowdSec Local API it could
happen that an error occurred, for example when the bouncer
was not authenticated or CrowdSec was not reachable. Before
the changes Caddy would simply stop because of a call to
log.Fatalf in the go-cs-bouncer implementation.

This commit includes a copy of the go-cs-bouncer without
the fatal calls, resulting in more control over what happens
when an error occurs when reaching out for the CrowdSec
Local API. An option to fail hard (or not) was added to the
CrowdSec app configuration too. It defaults to false.

Connection errors will always be logged, no matter the
configuration of failure mode.