Emit warning log when (likely) using low event throttle interval
When using the CrowdSec module with streaming mode enabled, together
with the Caddy Docker Proxy module, errors could occur when events
were not throttled sufficiently. The Caddy configuration reloads would
result in errors while trying to shut down the old server. This PR
adds a log to the CrowdSec provisioning process, indicating this is
(likely) the case, so that users can change it.

The logic only checks for the `CADDY_DOCKER_EVENT_THROTTLE_INTERVAL`
environment variable to be at least 2 seconds. It is also possible
to configure the Docker Proxy module using flags, but there's no way
the CrowdSec module can know the value that's provided through that.